python/requirements.txt
pyzmq>=25.1.0
redis>=5.0.0
uvloop>=0.19.0
